feat(core): replace per-shard (shard, clk) memory ordering with a global monotonic clk
MemoryRecord and every event/column that carried a memory-access
timestamp used a (shard, clk) pair, with clk resetting to zero at
each shard boundary. Replace it with a single u64 clk that never
resets across the whole execution, matching how MemoryLocalChip's
cross-shard GlobalChip accumulator needs to compare two accesses that
may land in different shards -- shard alone isn't enough there, since
it's the record's *previous* shard being compared against, not the
current row's own (already-constant) shard.

Executor/CoreVM/state.rs: clk becomes u64 and stops resetting per
shard; ExecutionState/CoreVM gain initial_timestamp, the clk value
when the current shard started, used both for "cycles consumed this
shard" (replacing the old reset-based count) and for detecting a
record's first touch within the current shard.

CpuState (adapter/state.rs) gains a clk_high limb alongside its
existing 28-bit low window (clk_16/8/4bit_limb) and range-checks it
the same way it already range-checks shard. clk_expr splits into
clk_high_expr/clk_low_expr; eval_state_chain and the hypercube
send_state/receive_state interaction grow a clk_high slot. Ordinary
per-access memory consistency (air/memory.rs, memory/consistency/*)
compares clk_high before clk_low, mirroring the old shard-then-clk
comparison. MemoryLocalChip/MemoryGlobalChip and SyscallChip's
cross-shard Global-kind linkage key off clk_high/clk_low instead of
shard, since the full clk value alone already uniquely identifies a
point in the whole execution.

SplicingVM cuts a shard early whenever continuing would cross a
1 << 28 clk_high window boundary, so clk_high stays constant across
every row of a shard's own trace -- no bump chip, no cross-row
transition logic, no new interaction kind needed for it.

PublicValues gains one new clk_high field; initial_timestamp/
last_timestamp keep representing just the low 28 bits, since
clk_high is constant within a shard and doesn't need a start/end
pair the way pc does.

register.rs's eval_register_reader, and several chips' own extra
eval_memory_access calls (mul/divrem's HI register, addi/load_word/
store_word, misc/others), were the last remaining call sites still
passing the real shard value where eval_memory_access now expects
clk_high -- fixed to match.

Adds a fast local-interaction-balance regression check against real
ELF programs (mips::tests::debug_real_elf_local_interactions_balance)
and a real end-to-end proving/verification test that forces many
small shards to exercise the cross-shard path densely
(run_test_many_small_shards).

+++ crates/core/executor/src/events/instr.rs
@@ -13,7 +13,7 @@ pub struct AluEvent {
     /// The shard number.
     pub shard: u32,
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
     pub pc: u32,
     pub next_pc: u32,
     /// The opcode.
@@ -89,7 +89,7 @@ pub struct CompAluEvent {
     /// The shard number.
     pub shard: u32,
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
 
     pub pc: u32,
     pub next_pc: u32,
@@ -170,7 +170,7 @@ pub struct MemInstrEvent {
     /// The shard.
     pub shard: u32,
     /// The clk.
-    pub clk: u32,
+    pub clk: u64,
     /// The program counter.
     pub pc: u32,
     pub next_pc: u32,
@@ -202,7 +202,7 @@ impl MemInstrEvent {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         shard: u32,
-        clk: u32,
+        clk: u64,
         pc: u32,
         next_pc: u32,
         opcode: Opcode,
@@ -239,7 +239,7 @@ pub struct BranchEvent {
     /// The shard number.
     pub shard: u32,
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
     /// The program counter.
     pub pc: u32,
     /// The next program counter.
@@ -270,7 +270,7 @@ impl BranchEvent {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         shard: u32,
-        clk: u32,
+        clk: u64,
         pc: u32,
         next_pc: u32,
         next_next_pc: u32,
@@ -305,7 +305,7 @@ pub struct JumpEvent {
     /// The shard number.
     pub shard: u32,
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
     /// The program counter.
     pub pc: u32,
     /// The next program counter.
@@ -336,7 +336,7 @@ impl JumpEvent {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         shard: u32,
-        clk: u32,
+        clk: u64,
         pc: u32,
         next_pc: u32,
         next_next_pc: u32,
@@ -371,7 +371,7 @@ pub struct MiscEvent {
     /// The shard number.
     pub shard: u32,
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
     /// The program counter.
     pub pc: u32,
     pub next_pc: u32,
@@ -402,7 +402,7 @@ impl MiscEvent {
     #[must_use]
     #[allow(clippy::too_many_arguments)]
     pub fn new(
-        clk: u32,
+        clk: u64,
         shard: u32,
         pc: u32,
         next_pc: u32,
@@ -440,7 +440,7 @@ pub struct MovCondEvent {
     /// The shard number.
     pub shard: u32,
     /// The clock cycle.
-    pub clk: u32,
+    pub clk: u64,
     /// The program counter.
     pub pc: u32,
     pub next_pc: u32,
@@ -470,7 +470,7 @@ impl MovCondEvent {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
         shard: u32,
-        clk: u32,
+        clk: u64,
         pc: u32,
         next_pc: u32,
         opcode: Opcode,
